Renamed MarksOfGraceCDOverlay.java > MarksOfGraceCDInfoOverlay.java for consistency.
